refactor: step4 리뷰 반영
