+++ src/main/kotlin/racingCar/domain/car/RandomMovingEngine.kt
@@ -1,15 +1,16 @@
-package racing_car.domain.car
+package racingCar.domain.car
 
 class RandomMovingEngine(
     private val minRandomValue: Int = 0,
     private val maxRandomValue: Int = 9,
     private val moveThreshold: Int = 4,
     private val defaultMoveDistance: Int = 1
-) :
+) : Engine {
 
-    Engine {
+    private val availableRange: IntRange = (minRandomValue..maxRandomValue)
+
     override fun getMoveDistance(): Int {
-        if ((minRandomValue..maxRandomValue).random() >= moveThreshold) {
+        if (availableRange.random() >= moveThreshold) {
             return defaultMoveDistance
         }
         return 0
