--- src/main/kotlin/racing_car/domain/race/RaceResult.kt
@@ -1,5 +1,0 @@
-package racing_car.domain.race
-
-data class RaceResult(val roundResults: List<RaceRoundResult>) {
-    val roundCount: Int = roundResults.size
-}
