+++ src/main/kotlin/step2/ExpressionParser.kt
@@ -9,4 +9,3 @@ interface ExpressionParser {
 
     fun parse(expression: String): Expression
 }
-
