+++ src/test/kotlin/racingCar/domain/car/CarTest.kt
@@ -1,4 +1,4 @@
-package racing_car.domain.car
+package racingCar.domain.car
 
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.shouldBe
