+++ src/test/kotlin/racingCar/domain/car/FixedMovingEngine.kt
@@ -1,4 +1,4 @@
-package racing_car.domain.car
+package racingCar.domain.car
 
 class FixedMovingEngine(private val defaultMoveDistance: Int = 1) : Engine {
     override fun getMoveDistance(): Int = defaultMoveDistance
